+++ module/data/item/talent.mjs
@@ -1,3 +1,5 @@
+import { systemPath } from "../../constants.mjs";
+import enrichHTML from "../../utils/enrich-html.mjs";
 import AdvancementModel from "./advancement.mjs";
 
 /**
@@ -6,6 +8,8 @@ import AdvancementModel from "./advancement.mjs";
 export default class TalentModel extends AdvancementModel {
   /** @inheritdoc */
   static LOCALIZATION_PREFIXES = super.LOCALIZATION_PREFIXES.concat("MYTHCRAFT.Item.talent");
+
+  /* -------------------------------------------------- */
 
   /** @inheritdoc */
   static defineSchema() {
@@ -18,5 +22,42 @@ export default class TalentModel extends AdvancementModel {
     schema.incompatibilities = new fields.StringField({ required: true });
 
     return schema;
+  }
+
+  /* -------------------------------------------------- */
+
+  /**
+   * The full list of tags as a human-readable string.
+   * @type {string}
+   */
+  get tagList() {
+    const tags = this.tags.reduce((tagList, tag) => {
+      const tagInfo = mythcraft.CONFIG.talents.tags[tag];
+      if (tagInfo) tagList.push(game.i18n.localize(tagInfo.label));
+      return tagList;
+    }, []);
+
+    return game.i18n.getListFormatter({ type: "unit" }).format(tags);
+  }
+
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
+  async toEmbed(config, options = {}) {
+    const enriched = await enrichHTML(this.description.value, { ...options, relativeTo: this.parent });
+
+    const embed = document.createElement("div");
+    embed.classList.add("mythcraft", this.parent.type);
+
+    const content = await foundry.applications.handlebars.renderTemplate(systemPath("templates/item/embeds/talent.hbs"), {
+      config,
+      enriched,
+      system: this,
+      systemFields: this.schema.fields,
+    });
+
+    embed.insertAdjacentHTML("afterbegin", content);
+
+    return embed;
   }
 }